feat: change label to labelinput, add labels list and label in context and reducer
