+++ src/components/LabelInput/LabelInput.css
@@ -2,7 +2,7 @@
   background-color: var(--white-color);
   padding: 1rem;
   border-radius: 10px;
-  width: 30%;
+  width: 35rem;
 }
 
 .label-input {
@@ -15,8 +15,4 @@
 
 .fa-circle-plus {
   font-size: 1.5rem;
-}
-
-.label-item {
-  width: fit-content;
 }
